fixing orders

+++ xchange-bybit/src/main/java/org/knowm/xchange/bybit/BybitAuthenticated.java
@@ -13,6 +13,7 @@ import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.QueryParam;
 import jakarta.ws.rs.core.MediaType;
 import java.io.IOException;
+import java.math.BigDecimal;
 import org.knowm.xchange.bybit.dto.BybitResult;
 import org.knowm.xchange.bybit.dto.account.allcoins.BybitAllCoinsBalance;
 import org.knowm.xchange.bybit.dto.account.feerates.BybitFeeRates;
@@ -81,6 +82,6 @@ public interface BybitAuthenticated {
       @FormParam("symbol") String symbol,
       @FormParam("side") String side,
       @FormParam("orderType") String orderType,
-      @FormParam("qty") long qty)
+      @FormParam("qty") BigDecimal qty)
       throws IOException, BybitException;
 }
